Move get_hosted_content_file and get_component into record.py
